refactor metrics.go and rename metrics
