+++ cloud/disk_manager/internal/pkg/services/pools/storage/metrics/ya.make
@@ -2,7 +2,6 @@ GO_LIBRARY()
 
 SRCS(
     metrics.go
-    metrics_impl.go
 )
 
 END()
