+++ cloud/disk_manager/internal/pkg/services/pools/storage/storage_ydb_impl.go
@@ -20,7 +20,7 @@ func (s *storageYDB) findBaseDisks(
 	ids []string,
 ) (baseDisks []baseDisk, err error) {
 
-	defer s.metrics.StatCall("storageYDB/findBaseDisks")(&err)
+	defer s.metrics.StatCall("findBaseDisks")(&err)
 
 	if len(ids) == 0 {
 		return nil, nil
@@ -623,7 +623,7 @@ func (s *storageYDB) getPoolOrDefault(
 	zoneID string,
 ) (poolInstance pool, err error) {
 
-	defer s.metrics.StatCall("storageYDB/getPoolOrDefault")(&err)
+	defer s.metrics.StatCall("getPoolOrDefault")(&err)
 
 	res, err := tx.Execute(
 		ctx,
@@ -1930,7 +1930,7 @@ func (s *storageYDB) getPoolConfigs(
 	session *persistence.Session,
 ) (configs []poolConfig, err error) {
 
-	defer s.metrics.StatCall("storageYDB/getPoolConfigs")(&err)
+	defer s.metrics.StatCall("getPoolConfigs")(&err)
 
 	res, err := session.StreamExecuteRO(ctx, fmt.Sprintf(`
 		--!syntax_v1
@@ -1984,7 +1984,7 @@ func (s *storageYDB) getBaseDisksScheduling(
 	session *persistence.Session,
 ) (baseDisks []BaseDisk, err error) {
 
-	defer s.metrics.StatCall("storageYDB/getBaseDisksScheduling")(&err)
+	defer s.metrics.StatCall("getBaseDisksScheduling")(&err)
 
 	res, err := session.ExecuteRO(ctx, fmt.Sprintf(`
 		--!syntax_v1
@@ -2032,7 +2032,7 @@ func (s *storageYDB) takeBaseDisksToScheduleForPool(
 	config poolConfig,
 ) (baseDisks []BaseDisk, err error) {
 
-	defer s.metrics.StatCall("storageYDB/takeBaseDisksToScheduleForPool")(&err)
+	defer s.metrics.StatCall("takeBaseDisksToScheduleForPool")(&err)
 
 	tx, err := session.BeginRWTransaction(ctx)
 	if err != nil {
@@ -2179,7 +2179,7 @@ func (s *storageYDB) takeBaseDisksToSchedule(
 	session *persistence.Session,
 ) (baseDisks []BaseDisk, err error) {
 
-	defer s.metrics.StatCall("storageYDB/takeBaseDisksToSchedule")(&err)
+	defer s.metrics.StatCall("takeBaseDisksToSchedule")(&err)
 
 	configs, err := s.getPoolConfigs(ctx, session)
 	if err != nil {
