+++ cloud/disk_manager/pkg/app/controlplane.go
@@ -358,7 +358,8 @@ func initControlplane(
 ) (serve func() error, err error) {
 
 	logging.Info(ctx, "Initializing pool storage")
-	poolStorage, err := pools_storage.NewStorage(config.GetPoolsConfig(), db, metrics.NewEmptyRegistry())
+	poolsMetricsRegistry := mon.NewRegistry("pools")
+	poolStorage, err := pools_storage.NewStorage(config.GetPoolsConfig(), db, poolsMetricsRegistry)
 	if err != nil {
 		logging.Error(ctx, "Failed to initialize pool storage: %v", err)
 		return nil, err
